code: place added icons at center of viewport
When adding an icon, it will now be placed at the center of the current
viewport, then shift the viewport right by 32px, the width of an icon.
By this token, icons will appear as expected without shifting the view
excessively, but will not overlap when multiple icons are placed.